Refactor use_hook for Rc reuse
Hook state that is stored as Rc<RefCell<dyn Any>> can be casted back to
the original Rc<RefCell<InternalHook>> using ptr casting this allows the
use_hook to provide the Rc<RefCell<InternalHook>> to the runner. In the
case of use_ref this means that we don't need to wrap the state in
another Rc & RefCell.