Refactor listeners to use StaticEvent trait
`CustomEventHandler` trait has been changed to `StaticEvent` which better
reflects it's purpose - to statically describe an event.

`StaticEvent` can be implemented easily to the supported standard events which
allows for only one `Wrapper` and `Listener` implementation to be used to support
custom and standard events.

Added a doc comment for the lack of `web_sys::FormDataEvent` - this doc
comment for IDE users will be visible when using the type in the `html!`
macro.